Made `TextShimmerView` and moved `TextShimmer` over there
